requirements.txt
wmi
pywin32
pystray
Pillow
